requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
click==7.1.2
dj-database-url==0.5.0
Django==2.2.12
djangorestframework==3.11.0
gunicorn==20.0.4
pathspec==0.8.0
psycopg2-binary==2.8.5
pytz==2020.1
regex==2020.5.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.1.0
